requirements.txt
bs4>=4.0.0
imageio>=2.1.2
mechanize>=0.3.3
numpy>=1.13.3
Pillow>=4.3.0
socksipy-branch>=1.01
win_unicode_console>=0.5
demjson>=2.2.4
requests>=2.21.0
certifi>=2019.3.9
